chore(routes): update app/patient to / routes & URL redirection
