+++ edgar-app/src/components/pages/UnprotectedPage.tsx
@@ -21,7 +21,7 @@ const UnprotectedPage = ({ children }: { children: JSX.Element }): JSX.Element =
 		auth.checkToken().then((res) => {
 			if (res.status === 'success') {
 				if (router.query.redirect) void router.push(router.query.redirect as string);
-				else void router.push('/app/patient');
+				else void router.push('/');
 			} else authenticateHandler();
 		});
 	}, [router.isReady]);
